move typeExpr calculation to its own file
